requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.4
idna==3.2
praw==7.4.0
prawcore==2.3.0
pyTelegramBotAPI==3.8.3
python-dotenv==0.19.0
requests==2.26.0
update-checker==0.18.0
urllib3==1.26.6
websocket-client==1.2.1
